Add new default_controllers file
This commit adds a new version of the default_controllers.yaml file
(titled default_controllers_v2.yaml) that includes a joint effort
controller. This was in an effort to alter the controllers used to allow
dynamic planning without jittering of the robot.
TODO: figure out how to use the joint effort controller (just changing
the arg in panda_moveit.launch from
"position_joint_trajectory_controller" to
"effort_joint_trajectory_controller" was not successful. OR figure out
how to alter the current (likely PD) control that is used to change the
gains and see if that impacts the jittering behavior.